Added help argument to the pikachuface module

+++ modules/pikachuface.py
@@ -9,8 +9,12 @@ import sqlite3
 import discord
 import os.path
 from discord.ext import commands
-from config.public import DATABASE_NAME
 from io import BytesIO
+from discord import Embed
+from discord import Colour
+from config.public import DATABASE_NAME
+from config.public import MIN_INSANITY_LEVEL
+from config.public import MAX_INSANITY_LEVEL
 
 # Create logger
 log = logging.getLogger(__name__)
@@ -22,29 +26,46 @@ class PikachuFace:
     def __init__(self, bot):
         self.bot = bot
 
-    @commands.command(name='pf', description="A cute pikachu to react with :3")
+    @commands.command(name='pf', 
+                      aliases=['pikachu','pikachuface','surprisedpikachu'],
+                      brief='react with a surprised pikachu face',
+                      description="display a surprised pikachu face of intensity [int] or of type [type]",
+                      usage="[int | type]")
     async def pf_retrieve(self, ctx, attr = "0"):
 
         # Connection to the database
         conn = sqlite3.connect(DATABASE_NAME)
         
-        # Managing the command
-        try:
+        # Managing help argument
+        if attr == "help":
             cursor = conn.cursor()
-            name = "pf_" + ("il" if attr.isdigit() else "") + attr
-            cursor.execute("SELECT image_pika FROM pikachus WHERE name_pika=?", (name,))
-            response = cursor.fetchone()
-            if(response is None):
-                await ctx.send('⚠ **Error:** couldn\'t fetch any pikachu (argument ' + attr + ' not recognized) ⚠')
-            else:
-                response = response[0]
-                img = BytesIO(open(os.path.join("res", "pikachuface", response), 'rb').read())
-                await ctx.message.delete()
-                await ctx.send(file = discord.File(img.getvalue(), 'dummy.png'))
-        except sqlite3.OperationalError as e:
-            log.error("Command has failed : " + str(e))
-        finally:
-            conn.close()
+            cursor.execute("SELECT name_pika FROM pikachus")
+            response = cursor.fetchall()
+            arglist = [n[0] for n in response]
+
+            embed = Embed(color=Colour.gold(), title="😮 **Possible Arguments**")
+            embed.add_field(name="Number", value= str(MIN_INSANITY_LEVEL) + " up to " + str(MAX_INSANITY_LEVEL))
+            embed.add_field(name="Type", value="`"+'`\n`'.join(arglist[MAX_INSANITY_LEVEL+1:])+"`", inline=False)
+
+            await ctx.send(embed=embed)
+
+        # Managing command
+        else:
+            try:
+                cursor = conn.cursor()
+                cursor.execute("SELECT image_pika FROM pikachus WHERE name_pika=?", (attr,))
+                response = cursor.fetchone()
+                if(response is None):
+                    await ctx.send('⚠ **Error:** couldn\'t fetch any pikachu (argument ' + attr + ' not recognized) ⚠\n              You can find the possible arguments by writing `-pf help`')
+                else:
+                    response = response[0]
+                    img = BytesIO(open(os.path.join("res", "pikachuface", response), 'rb').read())
+                    await ctx.message.delete()
+                    await ctx.send(file = discord.File(img.getvalue(), 'dummy.png'))
+            except sqlite3.OperationalError as e:
+                log.error("Command has failed : " + str(e))
+            finally:
+                conn.close()
 
 # Setup
 def setup(bot):
